build: per-file warnings on own sources only, fix -Wswitch-enum compliance
CMake:
- Replace project_warnings interface library with set_source_files_properties
  on ZEPHYR_SOURCES / ARG_SOURCES so MOC-generated files never inherit our flags
- Extract ZEPHYR_SOURCES variable reused by qt_add_qml_module and the
  set_source_files_properties block
- Add -Wcovered-switch-default and -Wswitch-enum to Clang and GNU warning
  sets in both root and tests CMakeLists.txt for consistency
- Add per-file warnings inside add_zephyr_test macro (replaces target-level link)
- tst_qml already uses its own per-file set (TST_QML_SOURCES)

C++ warning fixes (clang-cl -Wall -Wextra -Wswitch-enum compliance):
- serialhandler: replace memcpy with std::bit_cast for type-safe deserialisation,
  switch cstring for <array>/<bit>/<algorithm>, enumerate all switch cases
  explicitly instead of default to satisfy -Wswitch-enum
- timeserieschartmodel: DataPoint::values changed from C array to
  std::array<qreal,9>, add size_t casts for indexed access; xMin/xMax
  cast to qreal explicitly
- sensorreadingmodel: static_cast<int> on qsizetype comparisons and
  beginInsertRows/beginRemoveRows arguments; use qsizetype loop vars in pruning
- thresholdmanager: single-exit getColorForSensor with result variable,
  explicit qreal casts for float threshold comparisons
- sensorreading: reorder struct fields largest-alignment first (QDateTime
  before ints) to match padding-optimal layout; move timestamp Q_PROPERTY
  to match
- ioworker/iothread/serialhandler/databasemanager: add override to destructors
- csvexporter/ioworker: reorder members for reduced struct padding

Tests:
- tst_timeserieschartmodel: use size_t loop index for std::array access
- tst_sensorreadingmodel: static_cast<float>(i) in withCoordinates calls
- tst_databasemanager: static_cast<double> for float-to-double comparison

+++ src/models/sensorreadingmodel.h
@@ -41,7 +41,7 @@ public:
     QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
     QHash<int, QByteArray> roleNames() const override;
 
-    int count() const { return m_readings.count(); }
+    int count() const { return static_cast<int>(m_readings.count()); }
 
     Q_INVOKABLE void loadFromDatabase(const QDateTime &start, const QDateTime &end);
     Q_INVOKABLE void clear();
